code clean up
Former-commit-id: ad8c1624fc1b5af54f4eb7d3eac77e852de2e393

+++ ConsoleApplication/DataBrowserModule/src/org/janelia/it/workstation/gui/browser/gui/listview/icongrid/AnnotatedImageButton.java
@@ -184,7 +184,7 @@ public abstract class AnnotatedImageButton<T,S> extends SelectablePanel implemen
 
     public void setTitle(String title, int maxWidth) {
         // Subtle font size scaling 
-        int fontSize = (int) Math.round((double) maxWidth * 0.005) + 10;
+        int fontSize = (int) Math.round(maxWidth * 0.005) + 10;
         Font titleLabelFont = new Font("Sans Serif", Font.PLAIN, fontSize);
         titleLabel.setFont(titleLabelFont);
         titleLabel.setPreferredSize(new Dimension(maxWidth, titleLabel.getFontMetrics(titleLabelFont).getHeight()));
@@ -198,7 +198,7 @@ public abstract class AnnotatedImageButton<T,S> extends SelectablePanel implemen
             return;
         }
         // Subtle font size scaling 
-        int fontSize = (int) Math.round((double) maxWidth * 0.003) + 10;
+        int fontSize = (int) Math.round(maxWidth * 0.003) + 10;
         Font titleLabelFont = new Font("Sans Serif", Font.PLAIN, fontSize);
         subtitleLabel.setFont(titleLabelFont);
         subtitleLabel.setPreferredSize(new Dimension(maxWidth, subtitleLabel.getFontMetrics(titleLabelFont).getHeight()));
